+++ ReadItWebServer/src/main/java/com/dis/readit/dtos/book/BookReviewDto.java
@@ -17,7 +17,7 @@ public class BookReviewDto implements Serializable {
 	private Integer bookId;
 
 	@JsonProperty("nbrStars")
-	private Integer nbrStars;
+	private Double nbrStars;
 
 	@JsonProperty("content")
 	private String content;
